moved editors into folders
